style: 폰 테스트를 색상에 따라 분리
